Add authorisation token to request.
