+++ service/src/main/java/com/commercetools/pspadapter/payone/mapping/klarna/KlarnaRequestFactory.java
@@ -2,9 +2,9 @@ package com.commercetools.pspadapter.payone.mapping.klarna;
 
 import com.commercetools.pspadapter.payone.config.PayoneConfig;
 import com.commercetools.pspadapter.payone.domain.ctp.PaymentWithCartLike;
+import com.commercetools.pspadapter.payone.domain.payone.model.common.ClearingType;
 import com.commercetools.pspadapter.payone.domain.payone.model.common.PayoneRequest;
 import com.commercetools.pspadapter.payone.domain.payone.model.common.PayoneRequestWithCart;
-import com.commercetools.pspadapter.payone.domain.payone.model.common.ClearingType;
 import com.commercetools.pspadapter.payone.domain.payone.model.common.StartSessionRequestWithCart;
 import com.commercetools.pspadapter.payone.domain.payone.model.klarna.KlarnaAuthorizationRequest;
 import com.commercetools.pspadapter.payone.domain.payone.model.klarna.KlarnaPreauthorizationRequest;
@@ -23,7 +23,6 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.Locale;
 import java.util.Optional;
-import java.util.function.BiFunction;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
@@ -35,6 +34,7 @@ import static java.util.Arrays.asList;
 
 public class KlarnaRequestFactory extends PayoneRequestFactory {
 
+    public static final String AUTHORIZATION_TOKEN = "authorization_token";
     @Nonnull
     private final CountryToLanguageMapper countryToLanguageMapper;
 
@@ -98,8 +98,9 @@ public class KlarnaRequestFactory extends PayoneRequestFactory {
      * @param paymentWithCartLike {@link PaymentWithCartLike} from which read the values
      * @return same {@code request} instance.
      */
-    protected PayoneRequest mapKlarnaMandatoryFields(@Nonnull PayoneRequest request,
+    protected <BKR extends PayoneRequestWithCart>  void mapKlarnaMandatoryFields(@Nonnull BKR request,
                                                      @Nonnull PaymentWithCartLike paymentWithCartLike) {
+         // authorisationToken  - must be mapped in paydata
         // fistsname, lastname, street, zip, city, country, email - must be mapped in mapBillingAddressToRequest
         // gender - in mapCustomerToRequest
         // language, amount, currency in mapFormPaymentWithCartLike
@@ -111,10 +112,14 @@ public class KlarnaRequestFactory extends PayoneRequestFactory {
 
         mapLanguageFromCountry(request, paymentWithCartLike.getCartLike());
 
-        return request;
     }
 
-    private static void mapKlarnaCustomFields(@Nonnull PayoneRequest request, @Nonnull CustomFields customFields) {
+    private static void mapKlarnaCustomFields(@Nonnull PayoneRequestWithCart request, @Nonnull CustomFields customFields) {
+
+        // Add authorisation token to paydata
+        mapCustomFieldIfSignificant(customFields.getFieldAsString(AUTHORIZATION_TOKEN),
+                token -> request.appendPaymentData(AUTHORIZATION_TOKEN, token));
+
         mapCustomFieldIfSignificant(customFields.getFieldAsString(IP_FIELD), request::setIp);
 
         mapCustomFieldIfSignificant(customFields.getFieldAsDate(BIRTHDAY_FIELD), request::setBirthday,
